CRUD

+++ EONIS_IT34_2020/EONIS_IT34_2020/Controllers/AdministratorController.cs
@@ -25,13 +25,26 @@ namespace EONIS_IT34_2020.Controllers
         }
 
         [HttpGet]
-        //[Authorize(Roles = "Admin")]
+        //[Authorize(Roles = "Administrator")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         //[ProducesResponseType(StatusCodes.Status401Unauthorized)]
         //[ProducesResponseType(StatusCodes.Status403Forbidden)]
         public ActionResult<List<AdministratorDto>> GetAdministrator(int page = 1, int pageSize = 10)
         {
+            /*
+             if (!HttpContext.User.Identity.IsAuthenticated)
+           {
+               return Unauthorized("Da biste izvršili operaciju, morate kreirati nalog!");
+           }
+           var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role && (c.Value == "Administrator" ));
+
+           if (roleClaim == null)
+           {
+               return Forbid();
+           }
+           */
+
             var administratori = administratorRepository.GetAdministrator();
 
             if (administratori == null || administratori.Count == 0)
@@ -54,7 +67,6 @@ namespace EONIS_IT34_2020.Controllers
             var itemsPerPage = administratoriDto.Skip((page - 1) * pageSize).Take(pageSize).ToList();
 
             return Ok(itemsPerPage);
-            //return Ok(administratoriDto);
         }
 
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -63,40 +75,52 @@ namespace EONIS_IT34_2020.Controllers
         //[ProducesResponseType(StatusCodes.Status401Unauthorized)]
         //[ProducesResponseType(StatusCodes.Status403Forbidden)]
         //[Authorize(Roles = "Admin")]
-        public ActionResult<AdministratorDto> GetAdministratorById(Guid Id_administrator) //dto?
+        public ActionResult<AdministratorDto> GetAdministratorById(Guid Id_administrator) 
         {
+            /*
+             if (!HttpContext.User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Da biste izvršili operaciju, morate kreirati nalog!");
+            }
+            var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role && (c.Value == "Administrator" || c.Value == "Korisnik"));
+
+            if (roleClaim == null)
+            {
+                return Forbid();
+            }
+             */
             var administrator = administratorRepository.GetAdministratorById(Id_administrator);
 
             if (administrator == null)
             {
-                return NotFound();
+                return NotFound("Administrator with the specified ID not found.");
             }
             return mapper.Map<AdministratorDto>(administrator);
         }
 
-        /*[ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [AllowAnonymous]
-        [HttpGet("{KorisnickoImeAdministratora}")]
-        public ActionResult<AdministratorDto> GetAdministratorByKorisnickoIme(string KorisnickoImeAdministratora)
+        [HttpGet("korisnickoIme/{KorisnickoImeAdministratora}")]
+        public ActionResult<AdministratorDto> GetAdministratorByKorisnickoIme(string korisnickoIme)
         {
-           
-            var administrator = administratorRepository.GetAdministratorByKorisnickoIme(KorisnickoImeAdministratora);
+
+            var administrator = administratorRepository.GetAdministratorByKorisnickoIme(korisnickoIme);
 
             if (administrator == null)
             {
-                return NotFound();
+                return NotFound($"Administrator with the specified KorisnickoIme {administrator.KorisnickoImeAdministratora} not found.");
             }
 
             return mapper.Map<AdministratorDto>(administrator);
-        }*/
+        }
 
-
+ 
         [HttpPost]
         [Consumes("application/json")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public ActionResult<AdministratorDto> CreateAdministrator([FromBody] AdministratorCreationDto administratorCreationDto) //izmeniti!
+        public ActionResult<AdministratorDto> CreateAdministrator([FromBody] AdministratorCreationDto administratorCreationDto) 
         {
 
             if(administratorCreationDto == null)
@@ -109,21 +133,9 @@ namespace EONIS_IT34_2020.Controllers
                 Administrator createdAdministrator = administratorRepository.CreateAdministrator(administratorCreationDto);
                 return mapper.Map<AdministratorDto>(createdAdministrator);
 
-                /*var administrator = mapper.Map<Administrator>(administratorCreationDto);
-                administratorRepository.CreateAdministrator(administrator);
-                administratorRepository.SaveChanges();
-
-                var administratorDto = mapper.Map<AdministratorDto>(administrator);
-
-                return Ok(administratorDto);*/
-
-                //--
-
                 /*Administrator administratorEntity = mapper.Map<Administrator>(administratorCreationDto);
                 administratorEntity.Id_administrator = Guid.NewGuid();
-                AdministratorDto confirmation = administratorRepository.CreateAdministrator(administratorEntity); // izmeniti
-
-                Console.WriteLine(mapper.Map<AdministratorDto>(confirmation));
+                AdministratorDto confirmation = administratorRepository.CreateAdministrator(administratorEntity);
 
                 return Ok(mapper.Map<AdministratorDto>(confirmation));*/
             }
@@ -134,48 +146,40 @@ namespace EONIS_IT34_2020.Controllers
         }
 
         [HttpPut]
-        //[Authorize(Roles = "Admin")]
+        //[Authorize(Roles = "Administrator")]
         [Consumes("application/json")]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public ActionResult<AdministratorDto> UpdateAdministrator(AdministratorUpdateDto administrator) // (Administrator administrator)
+        //[ProducesResponseType(StatusCodes.Status403Forbidden)]
+        //[ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public ActionResult<AdministratorDto> UpdateAdministrator(AdministratorUpdateDto administratorUpdateDto) 
         {
             try
             {
-                var oldAdministrator = administratorRepository.GetAdministratorById(administrator.Id_administrator);
-                if(oldAdministrator == null)
-                {
-                    return NotFound();
-                }
-                Administrator administratorEntity = mapper.Map<Administrator>(administrator);
-                mapper.Map(administratorEntity, oldAdministrator);
-
-                oldAdministrator.ImeAdministratora = administrator.ImeAdministratora;
-                // dopuniti ostalo
-
-
-                /*if (!string.IsNullOrEmpty(korisnik.Lozinka))
-               {
-                   korisnik.Lozinka = BCrypt.Net.BCrypt.HashPassword(korisnik.Lozinka);
-               }*/
-
-                administratorRepository.SaveChanges();
-               
-                return Ok(mapper.Map<AdministratorDto>(administrator));
-
                 /*
-                 Administrator UpdateAdministrator(AdministratorUpdateDto administrator);
+                 if (!HttpContext.User.Identity.IsAuthenticated)
+               {
+                   return Unauthorized("Da biste izvršili operaciju, morate kreirati nalog!");
+               }
+               var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role && (c.Value == "Administrator"));
 
+               if (roleClaim == null)
+               {
+                   return Forbid(); // "You don't have permission to update administrator."
+               } 
+                */
 
-                var updatedAdministrator = administratorRepository.UpdateAdministrator(administrator);
+                Administrator administrator = mapper.Map<Administrator>(administratorUpdateDto);
+                var updatedAdministrator = administratorRepository.UpdateAdministrator(administratorUpdateDto);
+
                 AdministratorDto updatedAdministratorDto = mapper.Map<AdministratorDto>(updatedAdministrator);
+
                 return Ok(updatedAdministratorDto);
-                 */
             }
             catch (KeyNotFoundException)
             {
-                return NotFound();
+                return NotFound("Administrator with the specified ID not found.");
             }
             catch (Exception e)
             {
@@ -187,20 +191,30 @@ namespace EONIS_IT34_2020.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [HttpDelete("{Id_administrator}")]
-        //[Authorize(Roles = "Admin")]
+        //[Authorize(Roles = "Administrator")]
         public IActionResult DeleteAdministrator(Guid Id_administrator)
         {
             try
             {
+                /*if (!HttpContext.User.Identity.IsAuthenticated)
+                {
+                    return Unauthorized("Da biste izvršili operaciju, morate kreirati nalog!");
+                }
+                var roleClaim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role && (c.Value == "Administrator"));
+
+                if (roleClaim == null)
+                {
+                    return Forbid();
+                }*/
                 Administrator administrator = administratorRepository.GetAdministratorById(Id_administrator);
                 if (administrator == null)
                 {
-                    return NotFound();
+                    return NotFound("Administrator with the specified ID not found.");
                 }
 
                 administratorRepository.DeleteAdministrator(Id_administrator);
                 administratorRepository.SaveChanges();
-                return StatusCode(StatusCodes.Status204NoContent);
+                return StatusCode(StatusCodes.Status204NoContent, "\"Administrator with the specified ID successfully deleted.\"");
             }
             catch
             {
